Clang code formatting example. V1.
- Add .clang file and try it on one file

+++ C++/Yandex/src/0_white_belt/week_1/week_1.cpp
@@ -101,7 +101,7 @@ void FindSecondF() {
 			index = -1;
 		}
 		else if (str[i] == 'f' && cur_rep_numb == 1) {
-			index = i;
+			index = static_cast<int>(i);
 			break;
 		}
 	}
